HireMindAI/requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
requests==2.31.0
PyPDF2==3.0.1
groq==0.5.0
flask-cors==6.0.2